Fixed varying name length problem
for players that had a suffix, like OBJ, previously the Jr would end up
in the team name slot. Now that is fixed.

Discovered that table read ends after 50th row due to the insertion of
an ad. Trying to figure out how to get around that with the next()
function